requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
idna==3.3
Levenshtein==0.21.1
oauthlib==3.2.2
Pillow==9.3.0
python-Levenshtein==0.21.1
rapidfuzz==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
thefuzz==0.19.0
twitter==1.19.3
urllib3==1.26.8